+++ docs/whats_new.rst
@@ -22,6 +22,11 @@ Current 1.2 (dev)
 
 Enhancements
 ~~~~~~~~~~~~
+- Improving the docstring for :class:`braindecode.models.EEGNetv4`  (:gh:`768` by `Bruno Aristimunha`_)
+
+
+API changes
+~~~~~~~~~~~
 
 Bugs
 ~~~~